+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## 1.20.923.1
 
-* SCLabelPolicy
-  * Initial release
 * IntuneDeviceCategory
   * Initial Release;
+* SPOSite
+  * Fixed an issue where updating a site's properties
+    would throw an error complaining about the object
+    not being in a correct state.
 * DEPENDENCIES
+  * AADPreview Updated to 2.0.2.117;
+  * ExchangeOnlineManagement Updated to 2.0.3;
   * MSCloudLoginAssistant Updated to 1.0.40;
   * ReverseDSC Updated to 2.0.0.7;
 * Misc